Clean up constants
