+++ src/commonMain/kotlin/choliver/nespot/Constants.kt
@@ -14,6 +14,7 @@ const val TILE_SIZE = 8
 const val DOTS_PER_CYCLE = 3
 
 const val NAMETABLE_SIZE = 1024
+const val RAM_SIZE = 2048
 const val VRAM_SIZE = 2048
 const val CHR_SIZE = 8192
 const val PRG_RAM_SIZE = 8192
