+++ src/commonMain/kotlin/choliver/nespot/nes/Nes.kt
@@ -1,11 +1,11 @@
 package choliver.nespot.nes
 
 import choliver.nespot.CPU_FREQ_HZ
+import choliver.nespot.RAM_SIZE
 import choliver.nespot.apu.Apu
 import choliver.nespot.cartridge.Cartridge
 import choliver.nespot.cartridge.Rom
 import choliver.nespot.common.Address
-import choliver.nespot.common.Data
 import choliver.nespot.cpu.Cpu
 import choliver.nespot.cpu.Cpu.Companion.INTERRUPT_IRQ
 import choliver.nespot.cpu.Cpu.Companion.INTERRUPT_NMI
@@ -17,7 +17,7 @@ class Nes(
   rom: Rom,
   videoSink: VideoSink = object : VideoSink {},
   audioSink: AudioSink = object : AudioSink {},
-  private val onStore: ((Address, Data) -> Unit)? = null
+  private val intercept: ((Memory) -> Memory)? = null
 ) {
   private var steps = 0
 
@@ -29,7 +29,7 @@ class Nes(
     audioSink = audioSink
   )
 
-  private val cpuRam = Ram(CPU_RAM_SIZE)
+  private val cpuRam = Ram(RAM_SIZE)
 
   private val ppu = Ppu(
     memory = cartridge.chr,
@@ -47,7 +47,7 @@ class Nes(
   )
 
   private val cpu = Cpu(
-    maybeIntercept(cpuMapper),
+    intercept?.invoke(cpuMapper) ?: cpuMapper,
     pollInterrupts = ::pollInterrupts
   )
 
@@ -61,19 +61,6 @@ class Nes(
 
   private fun pollInterrupts() = (if (apu.irq || cartridge.irq) INTERRUPT_IRQ else 0) or (if (ppu.vbl) INTERRUPT_NMI else 0)
 
-  private fun maybeIntercept(memory: Memory) = if (onStore != null) {
-    val onStore = onStore
-    object : Memory {
-      override fun get(addr: Address) = memory[addr]
-      override fun set(addr: Address, data: Data) {
-        memory[addr] = data
-        onStore(addr, data)
-      }
-    }
-  } else {
-    memory
-  }
-
   inner class Diagnostics internal constructor() {
     val cpu = this@Nes.cpu.diagnostics
     val ppu = this@Nes.ppu.diagnostics
@@ -85,16 +72,4 @@ class Nes(
   val persistentRam = cartridge.persistentRam
 
   val diagnostics = Diagnostics()
-
-  // TODO - consolidate all the constants
-  companion object {
-    const val CPU_RAM_SIZE = 2048
-
-    const val ADDR_OAMDATA: Address = 0x2004
-    const val ADDR_OAMDMA: Address = 0x4014
-    const val ADDR_APU_STATUS: Address = 0x4015
-    const val ADDR_JOYPADS: Address = 0x4016
-    const val ADDR_JOYPAD1: Address = 0x4016
-    const val ADDR_JOYPAD2: Address = 0x4017
-  }
 }
