+++ src/jvmTest/kotlin/choliver/nespot/nes/CpuMapperTest.kt
@@ -2,11 +2,11 @@ package choliver.nespot.nes
 
 import choliver.nespot.apu.Apu
 import choliver.nespot.memory.Memory
-import choliver.nespot.nes.Nes.Companion.ADDR_APU_STATUS
-import choliver.nespot.nes.Nes.Companion.ADDR_JOYPAD1
-import choliver.nespot.nes.Nes.Companion.ADDR_JOYPAD2
-import choliver.nespot.nes.Nes.Companion.ADDR_JOYPADS
-import choliver.nespot.nes.Nes.Companion.ADDR_OAMDMA
+import choliver.nespot.nes.CpuMapper.Companion.ADDR_APU_STATUS
+import choliver.nespot.nes.CpuMapper.Companion.ADDR_JOYPAD1
+import choliver.nespot.nes.CpuMapper.Companion.ADDR_JOYPAD2
+import choliver.nespot.nes.CpuMapper.Companion.ADDR_JOYPADS
+import choliver.nespot.nes.CpuMapper.Companion.ADDR_OAMDMA
 import choliver.nespot.ppu.Ppu
 import choliver.nespot.ppu.Ppu.Companion.REG_OAMDATA
 import com.nhaarman.mockitokotlin2.doReturn
